requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
scikit-learn==1.4.1.post1
numpy==1.26.4
python-multipart==0.0.9
torch
torchvision
Pillow
